package store split to logic and IStore interface - dynamic add node and sync blocks support!
let store support database, cache or file storage interface.
